refactor: add IORedis and add BullMQ queue for search indexing
