+++ client/package.json
@@ -25,10 +25,12 @@
     "@types/ws": "^8.18.1",
     "@vercel/analytics": "^2.0.1",
     "bcrypt": "^6.0.0",
+    "bullmq": "^5.76.6",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "crypto": "^1.0.1",
     "framer-motion": "^12.29.2",
+    "ioredis": "^5.10.1",
     "jsonwebtoken": "^9.0.3",
     "lenis": "^1.3.17",
     "lucide-react": "^0.563.0",
