+++ client/src/app/api/send/code/check/route.ts
@@ -1,6 +1,6 @@
 import { ERRORS } from "@/constants/error-handling";
 import { getAuthUser } from "@/lib/auth";
-import client from "@/lib/client";
+import client from "@/lib/redis/client";
 import { NextResponse } from "next/server";
 
 export async function POST(req: Request) {
