+++ client/src/lib/queue/searchIndex.queue.ts
@@ -1,0 +1,8 @@
+import { Queue } from "bullmq"
+import connection from "../redis/connection"
+
+const searchIndexQueue = new Queue("search-index", { connection })
+
+await searchIndexQueue.setGlobalRateLimit(1, 1000)
+
+export default searchIndexQueue
