+++ client/src/utils/indexPost.ts
@@ -1,0 +1,6 @@
+
+
+
+export default function indexPost(postId: number) {
+  
+}
